sdks/python/requirements.txt
cryptography>=41.0.0
requests>=2.28.0
pyjwt>=2.6.0
flask>=2.2.0
python-dotenv>=1.0.0